Improved source code organization of plugins/variant
