+++ plugins/variants/src/VariantFeatureWidget/index.ts
@@ -1,4 +1,6 @@
+import { lazy } from 'react'
 import { ConfigurationSchema } from '@jbrowse/core/configuration'
+import WidgetType from '@jbrowse/core/pluggableElementTypes/WidgetType'
 import PluginManager from '@jbrowse/core/PluginManager'
 import { types } from 'mobx-state-tree'
 import { stateModelFactory as baseModelFactory } from '@jbrowse/core/BaseFeatureWidget'
@@ -13,5 +15,18 @@ export function stateModelFactory(pluginManager: PluginManager) {
       type: types.literal('VariantFeatureWidget'),
       descriptions: types.frozen(),
     }),
+  )
+}
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addWidgetType(
+    () =>
+      new WidgetType({
+        name: 'VariantFeatureWidget',
+        heading: 'Feature details',
+        configSchema,
+        stateModel: stateModelFactory(pluginManager),
+        ReactComponent: lazy(() => import('./VariantFeatureWidget')),
+      }),
   )
 }
