+++ plugins/variants/src/VcfAdapter/index.ts
@@ -1,1 +1,14 @@
-export { default as configSchema } from './configSchema'
+import PluginManager from '@jbrowse/core/PluginManager'
+import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'VcfAdapter',
+        configSchema,
+        getAdapterClass: () => import('./VcfAdapter').then(r => r.default),
+      }),
+  )
+}
