+++ plugins/variants/src/VcfTabixAdapter/index.ts
@@ -1,1 +1,16 @@
 export { default as configSchema } from './configSchema'
+
+import PluginManager from '@jbrowse/core/PluginManager'
+import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'VcfTabixAdapter',
+        configSchema,
+        getAdapterClass: () => import('./VcfTabixAdapter').then(r => r.default),
+      }),
+  )
+}
